cognipath-ai-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pandas==2.2.2
SQLAlchemy==2.0.29
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
google-generativeai==0.5.4
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
bcrypt==3.2.0
gunicorn==22.0.0